+++ test/runtests.jl
@@ -23,6 +23,7 @@ include("test_misc.jl")
 include("test_ops.jl")
 include("test_ad.jl")
 include("mixed_tensors.jl")
+include("test_einsum.jl")
 
 print_timer()
 println()